# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Todo List API
  description: A simple API for managing todo items
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Todos
    description: Operations related to todo items

paths:
  /api/todos:
    get:
      tags: [Todos]
      summary: Retrieve all todo items
      operationId: getTodos
      parameters:
        - in: query
          name: completed
          required: false
          description: Filter todos by completion status
          schema:
            type: boolean
      responses:
        '200':
          description: Todo items retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid completion filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Todos]
      summary: Create a new todo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  example: Buy groceries
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Title is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/todos/{id}:
    parameters:
      - in: path
        name: id
        required: true
        description: The todo's MongoDB ObjectId
        schema:
          type: string
    put:
      tags: [Todos]
      summary: Update a todo
      operationId: updateTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                done:
                  type: boolean
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid ID or update data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Todos]
      summary: Delete a todo
      operationId: deleteTodo
      responses:
        '200':
          description: Todo deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Todo:
      type: object
      required: [_id, title, done, createdAt, updatedAt]
      properties:
        _id:
          type: string
          example: 654a1f3e2b9c4a0012345678
        title:
          type: string
          example: Buy groceries
        done:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DeleteResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Todo deleted
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string